Check modify timestamp of unified kernel image
If you update your system before the backup copy could be created the
backup copy will be created from a UKI that was not successfully booted
at least once. To prevent this form happening the modify timestamp of
the UKI must be older that the last reboot or `ukibak` will not create a
copy.